Ai process

+++ backend/src/main/java/com/smartsquad/backend/controllers/PlayerController.java
@@ -1,6 +1,8 @@
 package com.smartsquad.backend.controllers;
 
+import com.smartsquad.backend.DTO.PlayerResponse;
 import com.smartsquad.backend.models.PlayerEntity;
+import com.smartsquad.backend.services.BalancingService;
 import com.smartsquad.backend.services.PlayerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -18,6 +20,7 @@ import java.util.List;
 public class PlayerController {
 
     private final PlayerService playerService;
+    private final BalancingService balancingService;
 
     /**
      * Get all players from the database.
@@ -28,13 +31,19 @@ public class PlayerController {
         return playerService.getAllPlayers();
     }
 
+
+    @PostMapping("/specific")
+    public List<PlayerEntity> getSpecificPlayers(@RequestBody List<String> names) {
+        return balancingService.findAllByNameIn(names);
+    }
+
     /**
      * Add a new player.
      * Access via: POST http://localhost:8080/api/players
      */
     @PostMapping
-    public PlayerEntity createPlayer(@RequestBody PlayerEntity player) {
-        return playerService.savePlayer(player);
+    public PlayerResponse createPlayer(@RequestBody PlayerEntity player) {
+         return playerService.savePlayer(player);
     }
 
     /**
@@ -42,7 +51,19 @@ public class PlayerController {
      * Access via: DELETE http://localhost:8080/api/players?name=messi
      */
     @DeleteMapping
-    public void deletePlayer(@RequestParam String name) {
+    public String deletePlayer(@RequestParam String name) {
+        if (name == null || name.isEmpty()) {
+           throw new IllegalArgumentException("Name cannot be null or empty");
+        }
+
         playerService.deletePlayer(name);
+        return name + " was deleted successfully";
     }
+
+    @DeleteMapping("/reset")
+    public String resetPlayers (){
+        playerService.deleteAllPlayers();
+        return "All players were deleted successfully";
+    }
+
 }
